api: fix type mismatches in geo handler scan targets
Change Commission from int to int64 to match argMax return type.
Change anchorPoints from int to uint64 to match count() return type.
The ClickHouse Go driver does not implicitly convert between integer
widths during Scan.

+++ api/handlers/geo_validators.go
@@ -19,7 +19,7 @@ type GeoValidatorItem struct {
 	Name        string  `json:"name"`
 	StakeSol    float64 `json:"stake_sol"`
 	StakePct    float64 `json:"stake_pct"`
-	Commission  int     `json:"commission"`
+	Commission  int64   `json:"commission"`
 	MetroCode   string  `json:"metro_code"`
 	CountryCode string  `json:"country_code"`
 	ASN         int64   `json:"asn"`
